`storage`: make `backlog_controller_config` properties `config::binding`s
The two use cases of the `backlog_controller` in the code presently
are the `compaction_controller` and the `upload_controller`. In both cases,
there are cluster properties associated with the PID values used in the
`backlog_controller`, which presently require a restart to take effect.

There is no strong reason for this to be the case. To enable making
these values changeable at run-time, alter the `backlog_controller_config`
and use of it in the `backlog_controller` to instead use `config::binding`s
around the raw value types.

While this does throw away some of the bare value type simplicity of the
`backlog_controller` design, the benefits and actual use cases of this
class in the codebase are a good motivation for making this change.