c/data-migrate/worker: fix a race condition
The worker loop fiber may be super quick and skip yielding on its
scheduling points. In this case, instead of spawning the loop fiber in
the background it effectively runs synchronously for a while. Sometimes
it resulted in getting the work done, setting promise value and
unmanaging the NTP together with the promise earlier than the future is
grabbed from this promise.

The fix is to grab the future before spawning the fiber that eventually
sets the promise value.