cl/task: changed the way how tasks are started
Previously logic about whether to start the task was distributed
between the task and the link (reconciler loop). Changed the
implementation to make the logic encapsulated in the `task` object. Now
the task has a pair of methods indicating if it should be started or
stopped.

In order to create a controller leader shard locked task there is a
`controller_linked_task` partial implementation of `task` interface.

Signed-off-by: Michał Maślanka <michal@redpanda.com>