archival: Make 'get_stream' reentrant
The get_stream lambda function is passed to the remote::upload_segment
method. The method invokes the lambda on every retry to get a new stream
(because input_stream in seastar is not seekable).

The problem is that the 'get_stream' lambda can only be called once. It
consumes the generated stream and sets the variable to nullopt. Next
call triggers exception. This commit returns the old behavior.

Signed-off-by: Evgeny Lazin <4lazin@gmail.com>