data_queue_test: add a fairness test
Data queue is intended to bound the memory and number of elements that
can persist in its underlying queue by blocking callers when full.

Data queue does not block fairly though. A large push may be blocked
while small pushes are allowed to bypass the waiter queue and join the
data queue immediately.

In the case of many small pushes 'cutting the line' the waiter queue
will be starved indefinitely.

This commit adds a test that demostrates this starving behavior. This
commit does NOT pass tests as a result.