app: do not add cloud topics app to deferred shutdown
The cloud topics app is constructed early so that other services can
acquire references to it. However, the way it was being constructed
added it to the deferred list before services that it depends on,
meaning that the shutdown order would be problematic if cloud topics
needed to interact with a service that had previously been shutdown
(e.g. partition manager in this case).

Instead we construct the instance directly, and rely on
application::shutdown to perform clean up in a reasonable order.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>