ct/l0: return control batches in frontend reader
When consuming transactional batches, there are control records emitted
by the rm stm into the local raft log. In order to properly allow
consumers to all offsets (for example the last offset in a partition
that is a commit marker) we need to read these offsets from the local
log. This means we need to remove our batch type filter and post filter
in the reader (as we need raft data batches with control markers as well
as placeholder batches). Then we update our logic to merge L0 data with
the local log by preserving these commit markers.

Note this change has no effect on reconcilation because it already uses
a transactional reader that filters out these control batches.

Other things discovered while debugging:
- these control batches are going to cause us to skip our cache as soon
  as it hits a control batch. We should fix this, probably the simplest
  thing is to add a cache lookup when hydrating from the local log after
  metadata is fetched.
- We should be careful when designing reconcilation that we don't spin
  trying to read the max offset of a log if it's a control batch that is
  filtered out. Presumably this isn't really an issue as I suspect we
  will fetch will polling similar to iceberg+wasm.