packaging: use a fixed install_path
We use fixed install_path in packaging now: this path is used to set
the interpreter and exists in the container. This path is

/opt/redpanda_installs/redpanda for redpanda
/opt/redpanda_installs/distributed_consumer_verifier for DVC

We still leave the `install_path` build setting in place for now because
we need to coordinate the changes here and in vtools, so we check this
in, then remove the vtools dependency, then remove the build setting.