tuner: Fix NIC auto addr detection
The net tuner would effectively only work automatically if the user
would be listening on "0.0.0.0".

This is very unlikely and hence it would never apply any `net` tuning
unless the interface was manually specified.

Fix the detection by actually checking the interfaces for the right IP
and not a hardcoded "0.0.0.0".

Fixes CORE-1865