direct_consumer/verifier/application monotonic fix
Weakens the non-monotonic status return to require last_fetched_offset
to be strictly less than the last seen last fetched offset.

Direct consumer allows for duplicates to be returned such as [0,50] and
then again [50]. Non-monotonic should only be flagged if the last
fetched offset actually decreases.

This fixes the detection logic to reflect this.