kafka/server: only disable strict max bytes for obligatory batch reads
Prior to this commit all partitions in the initial call to `fetch_ntps`
were allowed to exceed the max bytes configured in the log reader if it
was needed to read a single batch. Beyond not being required by the
Kafka protocol this also results in memory safety concerns. The fetch
path only reserves memory up to the configured max bytes. So every byte
read beyond that limit increases the likelihood of an OOM.