`ct/l1`: add `log_collector`
Abstract class for implementing a push based `log_collector` for registering
and deregistering CTPs/logs that are eligible for compaction on this broker to
the `compaction_scheduler`.

To decouple the `log_collector` from the `compaction_scheduler` and avoid
a circular dependency, instead of holding a `compaction_scheduler*` pointer,
callback functions are used instead.