`ct/l1`: move `compaction::sink` to `compaction::in_memory_sink`
This was a rough swing at a smoke test with the IO implementations
which uses an in-memory buffer to filter and create an L1 object.

While helpful for testing, the `compaction::sink` will be evolving
far past this point. Copy the current implementation to
`tests/in_memory_sink.{cc/h}`. `compaction/sink.{cc/h}` will be changed
in a future commit soon after.