ct: repurpose manager for leadership notifications
This PR repurposes the cloud_topics_manager to be really a centralized
abstraction to follow partition leadership for all cloud topics related
topic partitions. We have a number of places we do this notification
business with the cluster infrastructure. Let's just do it once, as I'm
about to add another component that follows the CTP leader, so I didn't
want do all the boilerplate for yet another use case. I'm sure we'll
have more going forward.

Future commits will switch the reconciler to this mechanism, as well as
the domain manager.

Additionally, we have to rename the L0 GC methods so it can be used
within ss::sharded (the stop method has to return a future).