ct/l1: create metastore topic on metastore::object_builder()
There currently isn't a good spot to create the metastore topic before
actually sending out RPCs to add objects. This meant that we could fail
to add objects to the object builder for a while, because the
replicated_object_builder needs the partition count of the metastore
topic to know how to partition objects.

This commit makes object_builder() an async method that creates the
metastore topic if it doesn't exist.