tests/rpk: tolerate `-1` in epoch field
It may sometimes not be guaranteed to have the up to date leader epoch
information in the partition metadata. The `-1` is valid as it indicates
that no leader epoch is provided. We should still return partition
description and if the test is interested in leader epoch it should
assert that separately.

Signed-off-by: Michał Maślanka <michal@redpanda.com>