k/batch_adapter: always set short_read flag when batch is incomplete
Previously when a batch buffer was shorter than the size of two fields
indicating the base offset and batch the `short_read` flag wasn't set.
This lead to a situation in which a reader continued and accessed
not initialized booleans triggering undefined behavior sanitizer. Now
that the flag is correctly set and boolean flags are recognized the
reader will correctly return an error or ignore partial batch if it is
configured to do so.

Signed-off-by: Michał Maślanka <michal@redpanda.com>