config: fix legacy_default functionality
Assume a property with 2 states, "A" (default) and "B".

On version "x" we decide to change default from "A" -> "B", using
legacy_default.

If a cluster on version "x-1" sets the property to "B", when it is
upgraded to "x", the property is set to "B" as part of application::hydrate_config.
Then, during the feature_table::set_original_version call, the value is
recognised as "default" and reverted to the old default ("A").

A flag to track whether the property value is set is added in this
commit. If a property has had it's value set at any point (weather to
the default value or otherwise), this marks it. Thus it knows to ignore
the legacy_default value, as it was explicitely added on the old
version.