iceberg/conversion: Map PB enum values to string, not int
Map protobuf enum values to their string name instead of int value. For
example, the enum

    enum state {
        Active = 0;
        Inactive = 1;
    }

previously would translate `Active = 0` to `0` and `Inactive = 1` to `1`,
but now will translate `Active = 0` to `"ACTIVE"` and `Inactive = 1` to
`"INACTIVE"` (note the capitalization, which comes from the PB
descriptor).

This is a breaking change. The type of columns translated from enums
will change from signed integer to string.

This also introduces a new error condition: previously, any integer
would be accepted as an enum value and translated to itself. Now, lookup
of the enum name will fail if the value isn't in the range of the enum.
This is a translation error because it means the value is invalid, at
least for the current schema. We could also handle it by using the enum
default, or by stringifying the value (`99` -> `"99"`).