k/server: track in flight requests per-connection
Tracks the version and in-flight-duration of the currently in-flight
requests, as well as the duration for which the connection has been
idle.

While some of this state is already tracked in
`client_protocol_state::_responses`, it does not currently track the
requests for the full lifetime of the request-response processing, so I
opted to track this state separately. In the future, we could consider
aligning these two states.