`metastore`: add `extent_metadata` to `compaction_offsets_response`
Compaction needs to be `extent` aware, as it is the unit of iteration/work for
a compaction job, and the `metastore` expects compaction updates to perfectly
bound the offset space of existing `extent`s already present in the `metastore`.
The `max_timestamp` is also returned, as it will be helpful for implementing
`min.compaction.lag.ms`.

Add the list of `extent`s` to the `compaction_offsets_response`.

The number of `extent`s may eventually grow too large for this to be included
in the RPC response to a compaction request. We may eventually want to think
about pagination or some other mechanism for compaction jobs to request
`X` number of `extent`s at a time, and refresh the list when necessary
during compaction.