ct: unify logging and reader creation in frontend
This unifies logging and creation of readers in the frontend to avoid
duplicated code. It also adds logging into the L1 reader constructor so
that we can maintain continutity in the logs as we hand off from the
frontend where we log extra context (e.g. LRO) before the reader is
created and then log reader identity once the reader is constructed.