ct: use implicit state machine in L1 reader
This commit removes the explicit states from the L1 reader and most of
the shared state (e.g. current object, current batches, ...) and encodes
the state machine directly in the coroutine logic.

This commit also addresses CORE-13950 which relates to the L1 reader
skipping offsets leading to gaps that produce incorrect fetch behavior.

The problem before was that reader would attempt to move the next offset
forward if no batches were read. However, the code which moved the
offset forward could not distinguish between no batches because the
reader read a full object vs no batches because the reader reach its
byte limit. In the later case the reader would increment the offset and
end up skipping batches when it should have returned what it had read
and the marked itself as end of stream.

This commit handles that case by identifying the over the limit scenario
and explicitly marking the reader as being at end of stream.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>