cl/rep: Added method to truncate shadow partition
When partition replicator receives a new batch, it will check the
offsets of the source and shadow partitions to see if the shadow
partition needs to be truncated.  If the source start offset is greater
than the shadow partitions source start AND it is less than the shadow
paritions HWM, then it will truncate the shadow partition.

Signed-off-by: Michael Boquard <michael@redpanda.com>