datalake: make remaining redpanda fields optional
Makes the remaining required fields in the redpanda struct optional.
This generally makes it easier to reason about evolving fields in this
struct (e.g. one strategy for handling breaking field changes is to
remove the broken field and then let Redpanda add it back with the new
type)..

NOTE: going from required to optional is not a breaking change in
Iceberg.