cloud_topics: fix reading transactional data from L1
When transactional data is read, there still might be transaction
markers in the rm_stm state. But since LRO has advanced, it's possible
the log is truncated, which can lead to translation errors for those
abort transactions. To fix this, we return a nullptr for the translator,
which tells the frontend that there are no aborted transactions that
need to be returned.