`storage`: remove transactional control batches during compaction
Similar to tombstone removal, this is a two step process:

1. During self compaction, all relevant transactional batches
have their transactional bit unset (analogous to the first pass in
tombstone removal in which all prior records with the same key as a
tombstone are removed) and the `tx_fence` control batch is removed.
2. The next time a control batch (a `raft_data` batch with the control
bit set) is encountered past the `delete.retention.ms` expiration horizon,
it can be removed (analogous to the second pass in tombstone removal,
though this horizon is set by the `self_compaction_timestamp`, not the
`clean_compact_timestamp`, for the differences in step 1.).

Of course, both tombstone and transactional control batch removal both
depend now on the `max_tombstone_remove_offset` set by coordinated compaction,
so the condition in step 2 also relies on the `segment`'s last/stable/dirty
offset being less than or equal to this offset.

Transactional control batch removal during compaction can be disabled via
the use of the cluster config `log_compaction_disable_tx_batch_removal` as
a safety hatch mechanism.