ct: plumb topic creation revision id into L0 write pipeline
The topic creation id is going to be plumbed through the write_request
into the batcher and used to generate the L0 object name. This commit
brings this revision id into the write pipeline where write requests are
created.

Since the revision id is an implementation detail from the POV of CT, we
cast it into cluster_epoch type before pushing into the CT universe.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>