rpk: add --if-not-exists flag to topic create
Adds --if-not-exists flag for idempotent topic creation, similar to SQL
CREATE TABLE IF NOT EXISTS. When set, existing topics return OK instead
of an error.

This command is used our getting started docs docker compose and
restarting docker compose makes it fail. We can of course wrap it in a
bash script which always checks first if a topic exists but having
support for this native in rpk is much cleaner.

(cherry picked from commit 955d4790242bf39f1ecb6e682f286e9e73074314)