treewide: ruff E721: type checks
Replace all use of '== type(...)' with 'is type(...)' for type checks.

This preserves the existing semantics. Arguably, some of these can be
replaced with isinstance() checks, but that is more likely to change
the semantics and I don't see any compelling reason to do so for these
limited cases.

Basically we silence the lint with the least invasive change.

Docs on this rule: https://docs.astral.sh/ruff/rules/type-comparison/