lsm/core: introduce file IDs
There are not that many kinds of files in an LSM tree. Really one kind
of data file: an SST, and a two metadata files: MANIFEST and CURRENT.
This commit introduced strongly typed handles for file IDs, which we
use to track relative age of files and use as a compact placeholder
for a filename, which can be generated from the file ID for each
filetype.

Latter commits will show how file IDs are used and stored when
`version_set` is introduced.