rptest: throw on unhealthy cluster
Always throw on unhealthy cloud cluster in _ensure_cluster_health, as
the old path is vestigial now that we are creating cloud clusters
before the run, and in any case we never want to exercise the recreation
workload.

Simplify the calling code as well since we always throw.

Finally, use more fine-grained exception types in the health checks so
we get a different PT signature for each one.