c/controller_stm: wait for all topic metadata to applied from snapshot
When controller stm snapshot is applied the sub-stm snapshots are
applied in parallel, there are only a few of stms that state is applied
first f.e. features table. Moved topic table state to be applied before
other components. Other parts of Redpanda like migrations or cluster
linking may depend on the topic table to be up to date.

Signed-off-by: Michał Maślanka <michal@redpanda.com>