c/partition: make sure that high watermark is offset translatable
It is possible that log is truncated before Raft visible index is
updated, in this situation translating the offset returned by
`cluster::hight_watermark` to kafka offset fails as the
`offset_translator` may no longer contain a range of not truncated high
watermark.

In order to fix the issue implemented the semantics in which an empty
log high watermark is equal to its start offset. Which translates to:
```
last_visible_index(empty_log) = start_offset - 1

```
Signed-off-by: Michał Maślanka <michal@redpanda.com>