[CORE-9236] rptest: stop datalake_verifier if all translated
We are seeing a failure of a test where it appears that we have
progressed up to the HWMs, but don't exist the wait(). This appears to
be because the verifier doesn't hold a lock in between checking for
translation completion and progress, which allows for a window for
progress to be made to get translated up to the HWMs.

This commit addresses this by having the progress function also check
for translation completion.