c/rm_stm: do not update LSO when begin/end tx is in progress
This commit introduces a new read/write lock that protects the LSO updates.
The lock is held for write when there is any operation that when applied
may influence calculation of LSO i.e. begin_tx and end transaction.

The lock is being acquired for read when `rm_stm` is being asked about
the LSO. If the lock is not acquired - a transaction operation is in
progress we fallback to previous LSO, otherwise the calculation is based
on the state that is guaranteed to be up to date.

Signed-off-by: Michał Maślanka <michal@redpanda.com>