direct_consumer/api_types.h
Add reasonable initializers to fetched_partition_data. This is not a
required correctness change, instead this is meant to preempt a writer
from forgetting to set a value when filling in the fields of
fetched_partition_data.

Adds are_offsets_equal to source_partition_offsets which will check if
tracked offsets are the same or different over time.