dt: fix await leader logic
When calling the await_stable_leader, the admin api will report the
cached state of the leadership of the partition on each node. After a
restart, this may be stale, but it will still be the same across every
node. When restart happens very quickly, it doesn't take into account
the new election that will start shortly after the restart, usually by
the previous leader. Until this election is finished, all list_offset
requests will return not_partition_leader. This may eat up some of the
available retry attempts of the kafka client. Because of the restarts,
there are also some few attempts burned at communicating sasl_credentials.
When it also happens to get a few broker_not_available errors thrown in
the mix, the retries run out and the test fails.

Switch to a more thorough wait condition to verify that a new leader has
been elected after the restart.