redpanda.py: clear id <-> index cache sometimes
We keep a cache of index to redpanda node id in redpanda.py, where index
is a 1-based index of the ClusterNode exposed by ducktape. Functions
like get_node_id(ClusterName) end up using this mapping to avoid
repeatedly querying redpanda for this information. However, this mapping
can be out of date if the redpanda node ids change, which happens in
tests that wipe redpanda and then restore it.

This leads to wrong node ids being exposed in functions like
RedpandaService.partitions(). This was recently exposed by the previous
commit, which throws on if it can't find the broker with the given
broker ID in partitions(), which was occurring due to above issues.

Fix this by clearing the index -> id mapping when we delete rp data
dirs, or when we clean the node, which is generally a precursor to
redpanda being restarted with (potentially) different broker ids. This
happens only for the node which was cleared.

This caused some RNOT tests to time out. The basic problem was that the
RNOT flow calls get_node_for_id, which ultimately causes `node_id` to be
called on nodes which have been killed/stopped, etc. Internally node_id
does many retries (30s timeout), and node_operations itself wraps that
with 10 retries, so this generally causes the test to time out one
way or another. Fix this by adding a check in node_id to see if the
node is started, and if not, throw immediately, and in get_node_by_id,
we do not call node_id on nodes that are not started.

Add a bit more logging in the node not found case, and in general around
the areas discussed above.