`storage`: prevent race between `disk_log_appender` and `remove_prefix_full_segments()`
We will be calling `remove_segment_permanently()` on what is potentially
the active segment in `remove_prefix_full_segments()`. If there is a concurrent
append occuring, we may get into a race with the `disk_log_appender` since we
return the units in `lock_holder` (i.e the segment's write lock) while the
future for `remove_segment_permanently()` is still unresolved.

In this case, we need to obtain units from `_segments_rolling_lock` to
prevent any concurrency issues, which manifest as either a thrown exception
due to gate closure, or an inability to add batches to a closed `segment`.

(cherry picked from commit e31ae0bdcac784752920d89fa81bf9f6f15b942f)