cl: Clamp RF to node count
If the source cluster happens to be larger than the shadow cluster, it
is possible that the RF of source topics may be too high.  In order to
permit shadowing, we will clamp the RF to the Shadow Cluster's node
count.

Signed-off-by: Michael Boquard <michael@redpanda.com>
(cherry picked from commit 872569e36be54e04fe1c7aa59cc99ca674770f7f)