redpanda: start sr, pp, audit if scheduling groups
In various fixture tests we pass in a fixed scheduling group to avoid
recreating them or having multiple of them during multi-node fixture
tests.

This injected scheduling groups, however, has also shortcut the
application::initialize function too eagerly, too early, and would avoid
initializing the pandaproxy, schema registry, audit logging.

This commit refactors the code to remove this implicit coupling.

The test behaviour is changed such that the audit logging client is now
always started.

However, the test behaviour is NOT changed in that pandaproxy and schema
registry is not started in these fixture tests, as enabling them
currently causes a range of assertion failures.