many_topics_test: unsafe decom flakey fix
Occasionally we are doubly unlucky.

First, test_decommission_node_unsafely will choose the 0th index seed
node as the node to restart.

Second, this node on restart will finish booting before it gets an
append entries from any other node in the cluster.

When both of these happen, the node declares itself the seed broker of a
new cluster and attempts to lead the remainder of the healthy nodes.

Resolution here is to provide the other brokers as seed nodes on restart
so the restarting node reaches out with a join node request.