cloud_io: Refresh credentials on TokenExpire error
The error is propagated from the cloud storage client as a dedicated
error code. The remote treats it as a non-retriable error which and
initiates a refresh.

There is a change to the test in the ntp_archiver_test test suite. The
existing test already triggers similar error so this commit just adds an
additional check that verifies that the token refresh is triggered.

Signed-off-by: Evgeny Lazin <4lazin@gmail.com>
(cherry picked from commit e179461ea760d061bc594e452fe46a53aaa0c984)