k/group_router: correctly pass require_stable flag in offset fetch
When offset fetch handler was changed to a new version supporting
querying multiple groups at a time w `require_stable` flag for the
request was lost. The default value for the flag is `false`. Even if
consumer fetched offsets with the `require_stable` flag set to true.

This small issue resulted in a very severe bug impacting
atomic-produce-consume workloads. Without the `require_stable` flag set
to `true` the consumer fetching offsets didn't know about an open
transaction in a group. This may lead to duplicates in the workload that
is promised to be 'exactly once'.

Signed-off-by: Michał Maślanka <michal@redpanda.com>
(cherry picked from commit f469df6aa78ba476f50d4f1311c47005897ee635)