lsm/io: split metadata and data persistence
There are use cases where we might want to put metadata persistence in
raft or elsewhere. So split metadata and data persistence explicitly
(instead of hacking around with filename parsing, etc). This allows us
to easily put the metadata on cloud storage as well in a way where we
can upload MANIFEST files with the raft term to prevent races with
zombie leaders.

Ultimately there is a lot of follow out of this change, but I think it's
a good one to split the metadata and SST file storage.