kafka/produce: futurize produce handler validation
currently there was an odd bit of validation that happened on the
partition shard, while some happened on the connection shard. To better
distribute work, I think it's better if more things like schema registry
validation was on the connection shard.

This also opens it up to futurize `validate_batch`, and do future
validation (like running a Wasm function) on the produce path.