test: handle rpk describe group before consumer offsets topic
We've seen describe group race with the creation of consumer offsets
topic. When this happens rpk describe will succeed, but it omits output
that depends on an accurate partition count for the consumer offsets
topic. This results in an altered output which the parsing logic doesn't
handle. Rather than making the parser smarter wait for the topic to
exist which should happen as a side effect and then retry.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>