rptest/rpk_tuner_test.py: fix whitespace issues
I edited this file which trimmed away trailing whitespace on some lines
which were part of a multi-line string. This broke the test, which does
an exact match. Since trailing whitespace is generally considered a rude
houseguest, we don't add it back, but rather do the comparison based on
a cleaned up version of the string, removing leading and trailing
whitespace and dropping empty lines.

Fixes CORE-14695.