raft/coordinated_recovery_throttle: allow units overuse
Throttler keeps a budget of units (bytes) which is refilled every tick
(once a second). It worked well with requests that are smaller than the
per-tick budget, but recent changes bumped max request size, so with
unfortunate combinations of config parameter values it can get stuck.

Allow the first request on the shard to overuse the allowance, restore
it upon tick accordingly. When restoring aim to even out the budgets on
shards if the shortage is global, redistribute according to demand
otherwise.

(cherry picked from commit aaef4bfb824e5ce404567ffaca1fd123dbea2906)