test: remove storage assumption from test
The two tests changed in this commit assumed S3 as a backend. When run
as cdt-on-azure they fail to connect to minio because that is the
configured S3 implementation when not run on actual S3 and of course it
fails because we don't startup minio.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>
(cherry picked from commit 56ad1d01905e9cfa464705eddb8184bf6bd4880b)