pre-commit-config: add proto to clang-format
clang-format formats .proto files, so we should add this to the
config (I had a CI run fail due to this omission).

Also, remove the files config from the clang-format section since it
is ANDed with types_or (despite the name, the or refers to what happens
inside the list of types_or only), so it's redundant and confusing to
have both.

(cherry picked from commit 89567ee2991cbb1d7feabd408f2312240497dc0b)