partition_balancer_*: implement auto decommission
Wires partition_autobalancing_node_autodecommission_time into partition
balancer.

This commit adds the basicmost implementation of auto decommissioning
which is based on the last seen from the perspective of the current
controller broker. This implementation will run into problems when
controller leadership changes.

In future commits, this will be changed for a coordinated approach where
the partition_balancer_planner will instead use the cluster health
report to seek the consent of a quorum of nodes before decommissioning a
broker.