`storage`: rename `may_have_transaction_batches`
This flag is only a marker for the existence of control batches within
a `segment`, _not_ any other sort of transactional data (e.g. raft data
batches with the transactional bit set).

We are going to have to indicate `segment`s as also having the transactional
raft data batches mentioned above with some sort of `may_have_transaction_data_batches`
flag in the future. To avoid confusion and be prescriptive about the flags name,
rename it to `may_have_transaction_control_batches`.