r/persisted_stm: do not load local snapshot if it is from 'future'
When `persisted_stm` is staring it uses the local snapshot to speed up
recovery and skip replying the whole log. It may happen that sometimes
the snapshot is not correctly removed from the kvstore in this case the
state machine may load the snapshot that was intended to be used for the
previos revision of the same NTP.

The local snapshot with incorrect offset may prevent the state machine
manager from applying the log entries to the state machines as its next
offset is incorrectly set.

This PR prevents the stm from loading the snapshot with offset that is
greater than current log dirty offset. This isn't a full fix of the
issue and will be accompanied with future work leading to changing the
way snaphots are stored in kvstore.

Signed-off-by: Michał Maślanka <michal@redpanda.com>
(cherry picked from commit 15be1203610171d347408557061d2f3b19a6423e)