`ct/l1`: add `compaction_epoch` to `partition_state`
Add a field which describes the current epoch of this partition's compacted
log. Effectively, it enumerates the number of logical changes made to the log's
underlying data, and will be incremented when a compaction or partial compaction
has been applied to this partition.

This is a useful field for optimistic concurrency control, and will be used
as a means to prevent a compaction job based on stale data from overwriting an
updated log.