direct_consumer/fetcher: vassert clarity fixes
Adds clarity to the meaning and implication of consistent partitions in
vasserts.

Namely, the code is written to check whether a partition is consistent
before operating on it, allowing us to skip most checks on iterators
to the end of a collection and nullopts from helper getting methods.

This was done to significantly cut down on invalid entry checks.