health_montior_backend: implement auto decom
Implements calculation of nodes past auto decom time in health montior
backend.

All nodes which are past the automatic decommission time will be added
to the auto_decommission_status field in the health report.

A node will be added if it was last seen prior to now - decom time
or if not seen ever this nodes uptime is prior to now - decom time