shadow_link: Have cluster link table hold shared_ptr
We have noticed that at scale, Shadow Linking has encountered some
issues when copying the state.  This is due to having a large number of
shadow topics.  This results in reactor stalls and issues with fail
over.  The solution is to eliminate the copies.

To do so, we are going to switch from holding a
`cluster_link::model::metadata` object in the table to a lw_shared_ptr
holding a const pointer to the metadata.  Upon update, the shared
pointer in the table will be replaced, however all current copies will
still be valid until their holders are notified of the update.

Signed-off-by: Michael Boquard <michael@redpanda.com>
(cherry picked from commit 258135f7a952617c5f75d3b53a521fff4af983de)