`rptest`: sort after appending `HEAD` in `upgrade_path_to_head()`
Occasionally we may mistag a release (e.g with the `rc` suffix), leading
to it not appearing in our `released_versions`. If we append `HEAD`
_after_ sorting the list, we may end up in a bad situation where the
upgrade path is reported as:

```
[DEBUG - 2025-12-15 14:19:31,569 - redpanda_installer - upgrade_path_to_head - lineno:587]: upgrade path to head from (24, 3, 5) is [(24, 3, 18), (25, 1, 12), (25, 2, 12), (25, 3, 2), (25, 2, 13)]
```

This above example was due to a mistag of version `v25.2.13` as `v25.2.13-rc1`.

Also, trim the list to `HEAD` in case it is not the highest major in
the list of `released_versions`. For example, for `v25.2.x` branch, the above
would be corrected from:

```
[(24, 3, 18), (25, 1, 12), (25, 2, 12), (25, 3, 2), (25, 2, 13)]
```

to:

```
[(24, 3, 18), (25, 1, 12), (25, 2, 12), (25, 2, 13), (25, 3, 2)]
```

and with the truncation to `HEAD`:

```
[(24, 3, 18), (25, 1, 12), (25, 2, 12), (25, 2, 13)]
```

(cherry picked from commit b69ad0d030a857781beb833eb87d7ea2310f56f7)