lsm: fix issue with level0 compactions and overlapping files
Now there is a very subtle issue in compaction of level0 files in that
if we expand the set of files due to an overlap, that file itself could
have caused an additional overlap, so we need to reset the search and
include the expanded space. This is due to level 0 allowing overlap of
the key space. When we did this reset, we were incrementing the counter
after we set i to 0. The proper fix, is to use the exact logic from
leveldb and only increment the counter after we select the file so the
resetting of `i` works.

This was found while writing a bunch of tests for the compaction picking
logic. So yay! for tests :)

LevelDB link:
https://github.com/google/leveldb/blob/ac691084fdc5546421a55b25e7653d450e5a25fb/db/version_set.cc#L498