utils: Add compiler_utils.h for misc compile hacks
It's a place to put REDPANDA_BEGIN_IGNORE_DEPRECATIONS and
REDPANDA_END_IGNORE_DEPRECATIONS which ignore deprecation warnings in
a region of code.

Sometimes we want to do this when we aren't going to remove the
deprecated usage immediate (often in upstream seastar) but need to
suppress the warning since we use -Werror.